- fix : 삭제 예외처리 작성

+++ src/main/java/com/example/newsfeedproject/service/PostService.java
@@ -68,7 +68,11 @@ public class PostService {
     public void deletePost(Long postId, String email) {
         User findUser = userRepository.findUserByEmail(email).orElseThrow(()
                 -> new ResponseStatusException(HttpStatus.NOT_FOUND, "이메일을 찾을 수 없음"));
-        findPostById(postId);
+        Post findPost = findPostById(postId);
+        findPost.getUser().getId();
+        if(!findPost.getUser().getId().equals(findUser.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "작성자만 게시글을 삭제할 수 있습니다");
+        }
         postRepository.deleteById(postId);
     }
